+++ src/python/pants/backend/docker/goals/publish.py
@@ -20,8 +20,8 @@ from pants.core.goals.publish import (
     PublishProcesses,
     PublishRequest,
 )
+from pants.core.util_rules.env_vars import environment_vars_subset
 from pants.engine.env_vars import EnvironmentVarsRequest
-from pants.engine.internals.platform_rules import environment_vars_subset
 from pants.engine.process import InteractiveProcess, Process
 from pants.engine.rules import collect_rules, implicitly, rule
