+++ src/python/pants/backend/docker/goals/run_image.py
@@ -16,9 +16,9 @@ from pants.backend.docker.target_types import (
 from pants.backend.docker.util_rules.docker_binary import DockerBinary
 from pants.core.goals.package import BuiltPackage, PackageFieldSet
 from pants.core.goals.run import RunFieldSet, RunInSandboxBehavior, RunRequest
+from pants.core.util_rules.env_vars import environment_vars_subset
 from pants.engine.env_vars import EnvironmentVarsRequest
 from pants.engine.internals.graph import resolve_target
-from pants.engine.internals.platform_rules import environment_vars_subset
 from pants.engine.rules import Get, collect_rules, concurrently, implicitly, rule
 from pants.engine.target import WrappedTargetRequest
