+++ src/python/pants/backend/tools/trivy/rules.py
@@ -8,10 +8,10 @@ from pants.backend.tools.trivy.subsystem import Trivy
 from pants.core.goals.resolves import ExportableTool
 from pants.core.util_rules import external_tool
 from pants.core.util_rules.config_files import find_config_file
+from pants.core.util_rules.env_vars import environment_vars_subset
 from pants.core.util_rules.external_tool import download_external_tool
 from pants.engine.env_vars import EnvironmentVarsRequest
 from pants.engine.internals.native_engine import Digest, MergeDigests
-from pants.engine.internals.platform_rules import environment_vars_subset
 from pants.engine.internals.selectors import concurrently
 from pants.engine.intrinsics import execute_process, merge_digests
 from pants.engine.platform import Platform
