+++ src/python/pants/core/register.py
@@ -53,6 +53,7 @@ from pants.core.util_rules import (
     adhoc_binaries,
     archive,
     config_files,
+    env_vars,
     external_tool,
     misc,
     source_files,
@@ -97,6 +98,7 @@ def rules():
         *archive.rules(),
         *build_configuration.rules(),
         *config_files.rules(),
+        *env_vars.rules(),
         *environments_rules.rules(),
         *external_tool.rules(),
         *git.rules(),
