+++ src/python/pants/jvm/test/junit.py
@@ -19,12 +19,12 @@ from pants.core.goals.test import (
     TestSubsystem,
 )
 from pants.core.target_types import FileSourceField
+from pants.core.util_rules.env_vars import environment_vars_subset
 from pants.core.util_rules.source_files import SourceFilesRequest, determine_source_files
 from pants.engine.addresses import Addresses
 from pants.engine.env_vars import EnvironmentVarsRequest
 from pants.engine.fs import DigestSubset, MergeDigests, PathGlobs, RemovePrefix
 from pants.engine.internals.graph import transitive_targets
-from pants.engine.internals.platform_rules import environment_vars_subset
 from pants.engine.intrinsics import digest_subset_to_digest, digest_to_snapshot, merge_digests
 from pants.engine.process import (
     InteractiveProcess,
